Payeezy (Payeezy v0.1.4)

Payeezy client library for Elixir. Right now this only supports GiftCardAdditional and GiftCardPurchase endpoints. More support will come as needed.

For api reference, please visit: https://developer.payeezy.com/payeezy-api/apis/

Functions

get_env(key)

@spec get_env(atom()) :: any()

Convenience function for retrieving payeezy specfic environment values, but will raise an exception if values are missing.

Example

iex> Payeezy.get_env(:random_value)
** (Payeezy.ConfigError) missing config for :random_value