Payeezy.GiftCard (Payeezy v0.1.4)

Payeezy gift card requests including balance_inquiry, purchase, and refund

Types

t()

@type t() :: %Payeezy.GiftCard{
  amount: String.t(),
  bank_message: String.t(),
  bank_resp_code: String.t(),
  correlation_id: String.t(),
  currency: String.t(),
  gateway_message: String.t(),
  gateway_resp_code: String.t(),
  method: String.t(),
  transaction_id: String.t(),
  transaction_status: String.t(),
  transaction_tag: String.t(),
  transaction_type: String.t(),
  validation_status: String.t(),
  valuelink: map()
}

Functions

activation(params)

@spec activation(map()) :: {:ok, t()} | {:error, any()}

Complete an activation for a given set of params

balance_inquiry(params)

@spec balance_inquiry(map()) :: {:ok, t()} | {:error, any()}

Complete a balance_inquiry for a given set of params

cashout(params)

@spec cashout(map()) :: {:ok, t()} | {:error, any()}

Complete a reload for a given set of params

deactivation(params)

@spec deactivation(map()) :: {:ok, t()} | {:error, any()}

Complete a deactivation for a given set of params

purchase(params)

@spec purchase(map()) :: {:ok, t()} | {:error, any()}

Complete a purchase for a given set of params

refund(params)

@spec refund(map()) :: {:ok, t()} | {:error, any()}

Complete a refund for a given set of params

reload(params)

@spec reload(map()) :: {:ok, t()} | {:error, any()}

Complete a reload for a given set of params

void(transaction_id, params)

@spec void(String.t(), map()) :: {:ok, t()} | {:error, any()}

Complete a void for a transaction_id and a given set of params