Payeezy.HTTP (Payeezy v0.1.4)

Base client for all server interaction, used by all endpoint specific modules. This request wrapper coordinates the remote server, headers, authorization and SSL options.

This uses HTTPoison.Base, so all of the typical HTTP verbs are available.

Using Payeezy.HTTP requires the presence of four config values:

apikey - Payeezy API key

apisecret - Payeezy API secret

token - Payeezy token

endpoint - Payeezy endpoint URL

All four must have values set or a Payeezy.ConfigError will be raised at runtime.

Types

body()

@type body() :: HTTPoison.Base.body()

headers()

@type headers() :: HTTPoison.Base.headers()

Functions

delete!(url, headers \\ [], options \\ [])

@spec delete!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a DELETE request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

delete(url, headers \\ [], options \\ [])

@spec delete(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a DELETE request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

get!(url, headers \\ [], options \\ [])

@spec get!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a GET request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

get(url, headers \\ [], options \\ [])

@spec get(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a GET request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

head!(url, headers \\ [], options \\ [])

@spec head!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a HEAD request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

head(url, headers \\ [], options \\ [])

@spec head(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a HEAD request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

options!(url, headers \\ [], options \\ [])

@spec options!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a OPTIONS request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

options(url, headers \\ [], options \\ [])

@spec options(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues an OPTIONS request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

patch!(url, body, headers \\ [], options \\ [])

@spec patch!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a PATCH request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

patch(url, body, headers \\ [], options \\ [])

@spec patch(binary(), any(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a PATCH request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

post!(url, body, headers \\ [], options \\ [])

@spec post!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a POST request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

post(url, body, headers \\ [], options \\ [])

@spec post(binary(), any(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a POST request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

process_headers(headers)

process_request_options(options)

process_response_chunk(chunk)

process_status_code(status_code)

put!(url, body \\ "", headers \\ [], options \\ [])

@spec put!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a PUT request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

put(url, body \\ "", headers \\ [], options \\ [])

@spec put(binary(), any(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a PUT request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

request!(method, url, body \\ "", headers \\ [], options \\ [])

@spec request!(atom(), binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues an HTTP request with the given method to the given url, raising an exception in case of failure.

request!/5 works exactly like request/5 but it returns just the response in case of a successful request, raising an exception in case the request fails.

request(method, url, body \\ "", headers \\ [], options \\ [])

@spec request(atom(), binary(), any(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues an HTTP request with the given method to the given url.

This function is usually used indirectly by get/3, post/4, put/4, etc

Args:

method - HTTP method as an atom (:get, :head, :post, :put, :delete, etc.)

url - target url as a binary string or char list

body - request body. See more below

headers - HTTP headers as an orddict (e.g., [{"Accept", "application/json"}])

options - Keyword list of options

Body:

binary, char list or an iolist

{:form, [{K, V}, ...]} - send a form url encoded

{:file, "/path/to/file"} - send a file

{:stream, enumerable} - lazily send a stream of binaries/charlists

Options:

:timeout - timeout to establish a connection, in milliseconds. Default is 8000

:recv_timeout - timeout used when receiving a connection. Default is 5000

:stream_to - a PID to stream the response to

:async - if given :once, will only stream one message at a time, requires call to stream_next

:proxy - a proxy to be used for the request; it can be a regular url or a {Host, Port} tuple, or a {:socks5, ProxyHost, ProxyPort} tuple *:proxy_auth- proxy authentication{User, Password}tuple *:socks5_user- socks5 username *:socks5_pass- socks5 password *:ssl- SSL options supported by thesslerlang module *:follow_redirect- a boolean that causes redirects to be followed *:max_redirect- an integer denoting the maximum number of redirects to follow *:params- an enumerable consisting of two-item tuples that will be appended to the url as query string parameters Timeouts can be an integer or:infinityThis function returnsorif the request is successful,` otherwise. ## Examples request(:post, "https://my.website.com", "{\"foo\": 3}", [{"Accept", "application/json"}])

start()

Starts HTTPoison and its dependencies.

stream_next(resp)

@spec stream_next(HTTPoison.AsyncResponse.t()) ::
  {:ok, HTTPoison.AsyncResponse.t()} | {:error, HTTPoison.Error.t()}

Requests the next message to be streamed for a given HTTPoison.AsyncResponse.

See request!/5 for more detailed information.